Generate binding redirects automatically instead of setting them explicitly in app.config
Previously, binding redirects were defined for licensekey dependency,
but the problem of version collision can happen for other modules as well
(e.g. when there is a dependency on a new version of core and an old version
of pdfCleanup which in turn depends on an older version of core). In those
cases because we have strongly-typed references, a runtime exception might
be thrown. Instead of adding new hard-coded binding redirects and maintaining
them on each release, the solution with automatic generation of binding
redirects allows us to get rid of defininf binding redirects at all,
even the ones for licensekey. Automatic binding redirect functionality is
enabled by default from 4.5.1 .NET framework versions but can be explicitly
enabled in earlier versions, which is necessary in our case

DEVSIX-3259