Add Antlr4.Runtime dependency
Now Antlr4 is a NuGet dependency in pdfXFA while previously
it used to be embedded in the assembly and loaded dynamically.
Now Antl4 dependency is defined in the NuSpec file and thus
it is added automatically when users add a reference to NuGet package.
In this commit we emulate adding pdfXFA via NuGet manager and thus
we add Antlr4 explicitly

DEVSIX-3082