Add Antlr3.Runtime dependency
Now Antlr3 is a NuGet dependency in pdfXFA while previously
it used to be embedded in the assembly.
Now Antl3 dependency is defined in the NuSpec file and thus
it is added automatically when users add a reference to NuGet package.
In this commit we emulate adding pdfXFA via NuGet manager and thus
we add Antlr3 explicitly

DEVSIX-3082